♻ [filestorage] Move class `GitRepositoryObserver` to test
